Added ExampleQuery
