+++ JebraAzureFunctions/JebraAzureFunctions/Properties/PublishProfiles/JebraFunctions - Zip Deploy.pubxml
@@ -1,0 +1,18 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<!--
+https://go.microsoft.com/fwlink/?LinkID=208121. 
+-->
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <WebPublishMethod>ZipDeploy</WebPublishMethod>
+    <PublishProvider>AzureWebSite</PublishProvider>
+    <LastUsedBuildConfiguration>Release</LastUsedBuildConfiguration>
+    <LastUsedPlatform>Any CPU</LastUsedPlatform>
+    <SiteUrlToLaunchAfterPublish>http://jebrafunctions.azurewebsites.net</SiteUrlToLaunchAfterPublish>
+    <LaunchSiteAfterPublish>False</LaunchSiteAfterPublish>
+    <ResourceId>/subscriptions/52b922a0-3bd6-4233-a626-a1c2b5a469e1/resourceGroups/Project-Jebra/providers/Microsoft.Web/sites/JebraFunctions</ResourceId>
+    <UserName>$JebraFunctions</UserName>
+    <_SavePWD>True</_SavePWD>
+    <PublishUrl>https://jebrafunctions.scm.azurewebsites.net/</PublishUrl>
+  </PropertyGroup>
+</Project>
